refactor(ui): split ui.ts into modular components (P1)
