+++ engineering_quality_summary.md
@@ -1,0 +1,74 @@
+# 工程质量分析与重构总结
+
+本文档总结了近期对 `chatgpt-multimodal-exporter` 项目进行的工程质量分析及初步重构工作，并对后续的重构计划进行了优先级标注。
+
+## 1. 现状分析 (Engineering Analysis)
+
+在重构开始前，我们对代码库进行了全面的审查，发现以下关键问题：
+
+### 1.1 代码质量与类型安全
+*   **问题**: 项目中大量使用了 `// @ts-nocheck` 指令，导致 TypeScript 的类型检查失效。
+*   **影响**: 代码脆弱，难以维护，重构时极易引入回归错误。缺乏类型提示降低了开发效率。
+*   **状态**: **已解决** (见第 2 节)。
+
+### 1.2 架构与模块化
+*   **问题**: `src/ui.ts` 文件过于庞大（800+ 行），集成了 UI 渲染、事件处理、业务逻辑和状态管理。
+*   **影响**: 违反了单一职责原则，代码阅读困难，修改 UI 容易影响业务逻辑。
+*   **问题**: UI 构建依赖手动 DOM 操作（`U.ce`），缺乏声明式 UI 的清晰度。
+
+### 1.3 外部依赖与稳定性
+*   **问题**: 项目核心功能高度依赖 OpenAI 的内部未公开 API (`/backend-api/...`)。
+*   **影响**: 这是一个极高风险点。OpenAI 随时可能更改接口结构或鉴权方式，导致插件完全失效。
+
+### 1.4 错误处理
+*   **问题**: 错误处理较为简陋，多处直接使用 `alert()` 或 `console.error`，缺乏统一的错误捕获和用户友好的反馈机制。
+
+---
+
+## 2. 已完成的重构工作 (Phase 1: Type Safety)
+
+针对上述分析中“代码质量”这一最基础的问题，我们执行了第一阶段的重构，重点在于**类型安全**。
+
+*   **建立类型系统**: 创建了 `src/types.ts`，定义了 `Conversation`, `Message`, `Attachment`, `Project`, `Task` 等核心接口。
+*   **移除 `@ts-nocheck`**: 彻底移除了所有源文件（`main.ts`, `ui.ts`, `utils.ts`, `api.ts`, `files.ts`, `downloads.ts`, `conversations.ts`, `batchExport.ts`, `cred.ts`）中的忽略检查指令。
+*   **全量类型覆盖**: 为所有函数参数、返回值和关键变量添加了显式类型注解，修复了构建时的类型错误。
+*   **构建验证**: 确保 `npm run build` 在严格模式下通过。
+
+## 2.1 已完成的重构工作 (Phase 2: Maintainability & UI)
+
+针对“架构与模块化”问题，我们执行了第二阶段的重构，重点在于**UI 模块化**。
+
+*   **UI 拆分**: 将庞大的 `src/ui.ts` 拆分为多个职责单一的模块。
+    *   `src/ui/styles.ts`: 集中管理 CSS 样式。
+    *   `src/ui/miniEntry.ts`: 负责悬浮球入口逻辑。
+    *   `src/ui/dialogs/`: 包含 `BatchExportDialog` 和 `FilePreviewDialog` 等业务弹窗组件。
+*   **入口重构**: `src/ui.ts` 现仅作为 UI 初始化的入口点，不再包含具体业务逻辑。
+
+
+---
+
+## 3. 重构优先级建议 (Refactoring Roadmap)
+
+基于工程分析，后续的重构建议按以下优先级进行：
+
+### 🔴 优先级 P0：核心稳定性 (Critical)
+*   **任务**: **API 适配层抽象**
+*   **描述**: 将所有对 OpenAI 内部 API 的调用（在 `api.ts` 和 `conversations.ts` 中）封装在一个独立的适配器层中。
+*   **理由**: 鉴于内部 API 的不稳定性，建立适配层可以使我们在 API 变更时，只需修改适配器代码，而无需改动业务逻辑。同时应增加对 API 响应结构的运行时校验（如使用 `zod`），以便在 API 变更时快速失败并定位问题。
+
+
+
+### 🟡 优先级 P2：用户体验 (Medium)
+*   **任务**: **优化错误处理与交互**
+*   **描述**: 引入一个轻量级的 Toast 通知系统替代 `alert()`。在网络请求失败时提供重试机制和更友好的错误提示。
+*   **理由**: 提升插件的专业感和用户体验。
+
+### 🔵 优先级 P3：工程化 (Low)
+*   **任务**: **引入单元测试**
+*   **描述**: 为 `utils.ts`, `files.ts` (解析逻辑) 等纯函数模块添加单元测试（如使用 `vitest`）。
+*   **理由**: 保证核心工具函数的正确性，防止未来修改导致的基础逻辑错误。
+
+### ⚪ 优先级 P4：技术栈升级 (Optional)
+*   **任务**: **引入 UI 框架**
+*   **描述**: 考虑在 `ui.ts` 中引入 Preact 或类似轻量级框架替代手动 DOM 操作。
+*   **理由**: 虽然能极大提升 UI 开发体验，但考虑到作为油猴脚本的体积限制和现有代码的迁移成本，建议在完成 P1 拆分后再考虑。
